Implementing DomainEventJsonDeserializer and DomainEventMapping
Description:

- DomainEventJsonDeserializer.

 Is in charge of converting a string event to a DomainEvent.

- DomainEventMapping

It is a collaborator of DomainEventJsonDeserializer. It has the
information about the relation between event types and DomainEvents in a
map. This map has been constructed dinamically (using the DI) calling
the subscribedTo method for the subscribers. So, it treats subscribers
as the source of truth.
________________________________________________________________________

There are more work to do in the next commits:

- Rethink the types, a huge amount of types with any right now
- DomainEvent is receiving the name as param in the constructor but
probably it won't be needed. That would allow to only declare an
abstract method that will be implemented in each DomainEvent.